feat: move field component return with the same create team api and disable save team on cost exceed
