+++ handlers/cmps/setup.go
@@ -11,7 +11,6 @@ import (
 // prefix path '/cmps'
 func SetupComponentRoutes(r_cmps chi.Router, repo *repos.Repository) {
 	r_cmps.Get("/players-table", handlers.Make(HandlePlayersTable, repo))
-	r_cmps.Post("/fantasy/player-card", handlers.Make(HandleFantasyPlayersField, repo))
 	r_cmps.Route("/admin", func(r_cmps_admin chi.Router) {
 		SetupAdminComponentRoutes(r_cmps_admin, repo)
 	})
