+++ views/fantasy.templ
@@ -26,7 +26,7 @@ templ Fantasy(
 			<div class="max-w-5xl">
 				<form
 					id="fantasy_player_selection_form"
-					hx-post="/cmps/fantasy/player-card"
+					hx-post="/api/fantasy/teams"
 					hx-trigger="change,submit"
 					hx-target="#team_field"
 					hx-swap="outerHTML"
